Move extend_freeorion_interface to separate folder
